# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AEGIS QUANT AI - Quantitative Trading Research Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /market/candles/{symbol}/{timeframe}:
    get:
      operationId: getCandles
      tags: [market]
      summary: Get OHLCV candles for a symbol and timeframe
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: timeframe
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Candle data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candle"
  /market/indicators/{symbol}/{timeframe}:
    get:
      operationId: getIndicators
      tags: [market]
      summary: Get computed indicators for a symbol and timeframe
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: timeframe
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Indicator values
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Indicators"

  # ── STRATEGIES ───────────────────────────────────────────────────────────────
  /strategies:
    get:
      operationId: listStrategies
      tags: [strategies]
      summary: List all strategies
      responses:
        "200":
          description: List of strategies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Strategy"
    post:
      operationId: createStrategy
      tags: [strategies]
      summary: Create a new strategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyInput"
      responses:
        "201":
          description: Strategy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
  /strategies/{id}:
    get:
      operationId: getStrategy
      tags: [strategies]
      summary: Get a strategy by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Strategy details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
        "404":
          description: Not found
    patch:
      operationId: updateStrategy
      tags: [strategies]
      summary: Update a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyUpdate"
      responses:
        "200":
          description: Updated strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
    delete:
      operationId: deleteStrategy
      tags: [strategies]
      summary: Delete a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /strategies/{id}/backtest:
    get:
      operationId: getStrategyBacktests
      tags: [strategies]
      summary: Get backtest results for a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Backtest results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BacktestResult"

  # ── SIGNALS ──────────────────────────────────────────────────────────────────
  /signals:
    get:
      operationId: listSignals
      tags: [signals]
      summary: List trading signals
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: symbol
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Signal"
    post:
      operationId: createSignal
      tags: [signals]
      summary: Create a new signal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignalInput"
      responses:
        "201":
          description: Signal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"

  # ── TRADES ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Candle:
      type: object
      required: [id, symbol, timeframe, open, high, low, close, volume, timestamp]
      properties:
        id:
          type: integer
        symbol:
          type: string
        timeframe:
          type: string
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
        timestamp:
          type: string
    Indicators:
      type: object
      required: [symbol, timeframe]
      properties:
        symbol:
          type: string
        timeframe:
          type: string
        ema20:
          type: ["number", "null"]
        ema50:
          type: ["number", "null"]
        ema200:
          type: ["number", "null"]
        rsi:
          type: ["number", "null"]
        macd:
          type: ["number", "null"]
        macdSignal:
          type: ["number", "null"]
        atr:
          type: ["number", "null"]
        vwap:
          type: ["number", "null"]
        bollingerUpper:
          type: ["number", "null"]
        bollingerLower:
          type: ["number", "null"]
        adx:
          type: ["number", "null"]
        updatedAt:
          type: string

    # ── STRATEGIES ──────────────────────────────────────────────────────────────
    Strategy:
      type: object
      required: [id, name, version, active, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        version:
          type: integer
        description:
          type: ["string", "null"]
        rulesJson:
          type: ["string", "null"]
        active:
          type: boolean
        winRate:
          type: ["number", "null"]
        totalTrades:
          type: ["integer", "null"]
        profitFactor:
          type: ["number", "null"]
        createdAt:
          type: string
    StrategyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        rulesJson:
          type: string
        active:
          type: boolean
    StrategyUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        rulesJson:
          type: string
        active:
          type: boolean

    # ── SIGNALS ─────────────────────────────────────────────────────────────────
    Signal:
      type: object
      required: [id, symbol, signalType, confidence, createdAt]
      properties:
        id:
          type: integer
        symbol:
          type: string
        strategyId:
          type: ["integer", "null"]
        strategyName:
          type: ["string", "null"]
        signalType:
          type: string
          enum: [buy, sell, hold]
        confidence:
          type: number
        reason:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
    SignalInput:
      type: object
      required: [symbol, signalType, confidence]
      properties:
        symbol:
          type: string
        strategyId:
          type: integer
        signalType:
          type: string
          enum: [buy, sell, hold]
        confidence:
          type: number
        reason:
          type: string

    # ── TRADES ──────────────────────────────────────────────────────────────────
...
    BacktestResult:
      type: object
      required: [id, strategyId, startDate, endDate, totalTrades, wins, losses, winRate, profitFactor, drawdown, createdAt]
      properties:
        id:
          type: integer
        strategyId:
          type: integer
        strategyName:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        totalTrades:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        winRate:
          type: number
        profitFactor:
          type: number
        drawdown:
          type: number
        sharpeRatio:
          type: ["number", "null"]
        totalReturn:
          type: ["number", "null"]
        createdAt:
          type: string
...