# lib/api-zod/src/generated/types/backtestResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AEGIS QUANT AI - Quantitative Trading Research Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BacktestResult {
  id: number;
  strategyId: number;
  /** @nullable */
  strategyName?: string | null;
  startDate: string;
  endDate: string;
  totalTrades: number;
  wins: number;
  losses: number;
  winRate: number;
  profitFactor: number;
  drawdown: number;
  /** @nullable */
  sharpeRatio?: number | null;
  /** @nullable */
  totalReturn?: number | null;
  createdAt: string;
}