# lib/api-zod/src/generated/types/strategyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AEGIS QUANT AI - Quantitative Trading Research Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface StrategyInput {
  name: string;
  description?: string;
  rulesJson?: string;
  active?: boolean;
}